ejemplo2/Ejemplo/pyproject.toml
[tool.poetry]
name = "ejemplo"
version = "0.1.0"
description = ""
authors = ["alexndra1272 <alexandrag1272@gmail.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.12"
flask = "^3.0.0"
python-dotenv = "^1.0.0"
flask-migrate = "^4.0.5"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
